requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

deta==1.2.0
langchain==0.0.268
nest-asyncio==1.5.7
openai==0.27.8
pinecone-client==2.2.2
requests==2.31.0
sentence-transformers==2.2.2
streamlit==1.25.0
streamlit-authenticator==0.2.3
streamlit-custom-notification-box==0.1.2
streamlit-javascript==0.1.5
transformers==4.31.0
verify-email==2.4.3
zibal==1.1.0
